refactored components so API's go in as a prop style changes & bug fixes
