+++ pages/api/get-orders_demo.js
@@ -2,7 +2,7 @@ import { sql } from '@vercel/postgres';
 
 export default async function getOrders(req, res) {
   const orders = (
-    await sql`SELECT * FROM demo_orders ORDER BY id DESC LIMIT 3;`
+    await sql`SELECT * FROM demo_orders WHERE status != 'completed' ORDER BY id DESC;`
   ).rows;
   res.status(200).json(orders);
 }
