+++ styles/index.module.css
@@ -14,6 +14,25 @@
   height: 100vh;
 }
 
+.editWholesale {
+  display: flex;
+  max-width: 100%;
+  overflow-x: auto;
+}
+
+.parentFlex {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  width: 100%;
+  margin: auto;
+
+  @media (min-width: 768px) {
+    width: 100%;
+  }
+}
+
 .parent {
   display: flex;
   justify-content: center;
@@ -34,17 +53,10 @@
 }
 
 .dash {
-  opacity: 0.9;
   display: flex;
   justify-content: center;
   align-items: center;
   flex-direction: row;
-  width: 100%;
-  max-width: 80rem;
-  padding: 1rem;
-  border-radius: 0.5rem;
-  background-color: #7e53b6;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
 }
 
 .dash button {
@@ -59,7 +71,7 @@
   font-size: 16px; /* adjust as needed */
 }
 
-.dash:hover {
+.dash button:hover {
   opacity: 0.5;
 }
 
@@ -88,11 +100,7 @@
 }
 
 .infoCard {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  width: 80%;
+  width: 90%;
   max-width: 600px;
   margin: auto;
   padding: 1rem;
